Ditch the transform for tooltips
With the transform, the tooltip starts slightly over the trigger, and
then moves out of the way on hover. This results in the tooltip itself
briefly stealing hover focus in some browsers. The result is that the
tooltip will flash in and out of visibility when hovering in specific
locations.